+++ src/m_misc.c
@@ -910,8 +910,8 @@ default_t defaults[] = {
   {
     "show_berserk",
     (config_t *) &show_berserk, NULL,
-    {3}, {0,3}, number, ss_weap, wad_yes,
-    "Show Berserk availability (0 = None, 1 = Ammo, 2 = Arms, 3 = Both)"
+    {1}, {0,1}, number, ss_weap, wad_yes,
+    "1 to display Berserk pack when using the Fist, if available"
   },
 
   {